add SpectraXML
